+++ lib/screens/main_page/main_page.dart
@@ -22,6 +22,8 @@ import '../notification/notification_page.dart';
 import '../size_config.dart';
 import '../view_detai_match/view_detail_match.dart';
 import '../view_list_match/view_list_match.dart';
+import '../view_result_team/component/view_result_team_menu.dart';
+import '../view_result_team/view_result_team.dart';
 import 'widget/component_button.dart';
 
 class MainPage extends StatefulWidget {
@@ -218,19 +220,9 @@ class _MainPageState extends State<MainPage> {
                     bloc: GetIt.I.get<
                         ViewCompetitionMemberTaskBloc>()), // trang show Competition mà Member này có task ở trỏng
                 CompetitionPage(bloc: GetIt.I.get<CompetitionBloc>()),
-                NotificationPage(bloc: GetIt.I.get<NotificationBloc>()),
-                //ViewListMatchPage(bloc: GetIt.I.get<ViewListMatchBloc>()),
-                // ViewDetailMatchPage(bloc: GetIt.I.get<ViewDetailMatchBloc>()),
-                //ViewDetailMatchPage(),
-                // ViewCompetitionRoundResultPage(bloc: GetIt.I.get<CompetitionRoundBloc>()),
-                //ViewListCompetitionOfParticipantPage(),
+                //NotificationPage(bloc: GetIt.I.get<NotificationBloc>()),
+                ViewResultTeamPage(),
                 ProfilePage(bloc: GetIt.I.get<ProfileBloc>())
-                //ViewDetailActivityPage(),
-                // EventPage(bloc: GetIt.I.get<EventBloc>()),
-                // ViewCompetitionMemberTaskPage(
-                //     bloc: GetIt.I.get<ViewCompetitionMemberTaskBloc>()),
-                // ViewListTeamParticipantPage(
-                //     bloc: GetIt.I.get<ViewListTeamParticipantBloc>()),
               ],
             ),
             bottomNavigationBar: BottomAppBar(
